+++ chat21/CHANGELOG.md
@@ -1,5 +1,9 @@
 ﻿# CHANGELOG
 
+## v.: 1.0.6 b.: 33
+- updated attachments metadata
+- update firebase dependencies resolution (from 11.6.0 to 11.8.0)
+
 ## v.: 1.0.5 b.: 32
 - show user id if username is not available
 - fixed autologin after signup - ChatLoginActivity